backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic>=2.8,<3.0
pydantic-settings>=2.9,<3.0
slowapi==0.1.9
SQLAlchemy>=2.0,<3.0
aiomysql>=0.2.0
